Refactor: Consolidate Go wrapper tests into main test file
The tests for the `io.Reader`/`io.Writer` adapters and the `DetectZxc` function have been moved from `zxc_io_test.go` into `zxc_test.go`. This change consolidates all Go wrapper tests into a single file, improving test discoverability and management.
